+++ src/lib/compiler/compileBootstrap.ts
@@ -22,7 +22,6 @@ interface InitialState {
   avatarFonts: PrecompiledAvatarData[][];
   engineFields: EngineFieldSchema[];
   engineFieldValues: EngineFieldValue[];
-  persistSceneSpriteSymbols: Record<string, string>;
 }
 
 const notDefine = (engineField: EngineFieldSchema) =>
@@ -39,7 +38,6 @@ export const compileScriptEngineInit = ({
   avatarFonts,
   engineFields,
   engineFieldValues,
-  persistSceneSpriteSymbols,
 }: InitialState) => `.include "vm.i"
 .include "macro.i"
 .include "data/game_globals.i"
@@ -93,14 +91,6 @@ ${engineFields
     return `        VM_SET_CONST_INT16      _${engineField.key}, ${value}`;
   })
   .join("\n")}
-
-${Object.keys(persistSceneSpriteSymbols)
-  .map(
-    (sceneType) =>
-      `        VM_SET_CONST      PLAYER_SPRITE_${sceneType}_BANK, ___bank_${persistSceneSpriteSymbols[sceneType]}\n` +
-      `        VM_SET_CONST      PLAYER_SPRITE_${sceneType}_DATA, _${persistSceneSpriteSymbols[sceneType]}\n`
-  )
-  .join("")}
 
         ; return from init routine
         VM_RET_FAR